test: 잘못된 경로의 중복 테스트 파일 삭제 및 정본 병합 (#171)
