+++ src/tests/unit/utils/test_redis_session.py
@@ -12,7 +12,6 @@ from src.utils.redis_session import RedisSessionStore  # noqa: E402
 
 @pytest.fixture
 def redis_store():
-    # Setup _get_client mock
     with patch("src.utils.redis_session._get_client", return_value=mock_redis.Redis.return_value):
         yield RedisSessionStore
 
@@ -45,12 +44,38 @@ def test_load_messages(redis_store):
     assert loaded_data == expected_data
 
 
-def test_delete_session(redis_store):
-    redis_store.delete_session("session_123")
-    assert mock_redis.Redis.return_value.delete.call_count >= 1
-
-
 def test_redis_retry_decorator_failure(redis_store):
     mock_redis.Redis.return_value.get.side_effect = Exception("Connection lost")
     result = redis_store.load_messages("test_sess")
     assert result == []
+
+
+@patch("src.utils.redis_session._get_client")
+def test_delete_session_success(mock_get_client):
+    mock_redis = MagicMock()
+    mock_get_client.return_value = mock_redis
+
+    test_session_id = "test_user_123"
+    expected_pattern = f"rag:session:{test_session_id}:*"
+    mock_keys_to_delete = [f"rag:session:{test_session_id}:messages", f"rag:session:{test_session_id}:docs"]
+
+    mock_redis.keys.return_value = mock_keys_to_delete
+
+    RedisSessionStore.delete_session(test_session_id)
+
+    mock_redis.keys.assert_called_once_with(expected_pattern)
+    mock_redis.delete.assert_called_once_with(*mock_keys_to_delete)
+
+
+@patch("src.utils.redis_session._get_client")
+def test_delete_session_no_keys(mock_get_client):
+    mock_redis = MagicMock()
+    mock_get_client.return_value = mock_redis
+
+    test_session_id = "empty_user_session"
+
+    mock_redis.keys.return_value = []
+
+    RedisSessionStore.delete_session(test_session_id)
+
+    mock_redis.delete.assert_not_called()
